requirements.txt
black==22.6.0
docker==6.0.0
isort==5.10.1
moto==4.0.1
pre-commit==2.20.0
pytest==7.1.2
pytest-cov==4.0.0
pytest-schema==0.1.1
pytest-xdist==3.1.0
python-dotenv==0.20.0
